Refactor move application gql mocks
