+++ apps/ui/test/applications/[id]/page3.test.tsx
@@ -2,12 +2,12 @@ import { type ApplicationPage3Query } from "@/gql/gql-types";
 import Page3 from "@/pages/applications/[id]/page3";
 import { render, within } from "@testing-library/react";
 import { vi, expect, test, describe } from "vitest";
+import { type CreateGraphQLMocksReturn } from "@/test/test.gql.utils";
 import {
   createMockApplicationFragment,
-  CreateMockApplicationFragmentProps,
-  type CreateGraphQLMocksReturn,
+  type CreateMockApplicationFragmentProps,
   createGraphQLApplicationIdMock,
-} from "@/test/test.gql.utils";
+} from "@/test/test.application.mocks";
 import userEvent from "@testing-library/user-event";
 import { getApplicationPath } from "@/modules/urls";
 import { MockedGraphQLProvider } from "@/test/test.react.utils";
